moved all config settings to single location in project source
closes issue #53
