+++ package.json
@@ -56,7 +56,6 @@
     "istanbul": "^0.4.1",
     "istanbul-instrumenter-loader": "^0.1.3",
     "jasmine-core": "^2.4.1",
-    "jquery": "^2.1.4",
     "karma": "^0.13.15",
     "karma-chrome-launcher": "^0.2.2",
     "karma-coverage": "^0.5.3",
